pyproject.toml
[tool.poetry]
name = "toolbox"
version = "0.1.0"
description = "toolbox by chatgpt"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pysrt = "^1.1.2"
mistune = "^2.0.5"
pydub = "^0.25.1"
pillow = "^9.5.0"
git-python = "^1.0.3"
huggingface-hub = "^0.14.1"
pytest = "^7.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
